Fix a crash that can happen on authentication if the view is not dismissed
If the web view loads m.box after the OAuth2 redirect, WebKit could crash
on inserting a nil value into a dictionary (history?). This seems to be a
WebKit bug, however loading the initial page in the web view via the
connection, as opposed to directly in the web view, avoids the situation.