Fix operation cancelling once and for all
The root cause of the issue was using main as the entry point to the
BoxAPIOperation instead of start.

This commit also removes the NSRunLoop and NSPort hacks used to
receive NSURLConnection delegate callbacks. The new approach is
to schedule a thread that runs forever and dispatch BoxAPIOperation
methods to this global thread.