Set the cookie acceptance policy appropriately
The authorization view controller now sets the cookie acceptance policy
to 'always' and resets it to the original value when done.
Also, the clearing of cookies is now done by maintaining pre-existing
cookies and deleting any that were created in the process of authorization
and login (because with Single Sign-On we cannot predict the domain
of cookies that will be created).
Some additional logging cleanup (no longer logging hex NSData).