Document that upload and download operations cannot be automatically retried
Upload and download SDK operations hold references to NSStreams for input and output.
NSStreams are not copyable so neither are these operations. The SDK relies on copying
operations to reenqueue them after an OAuth2 refresh event.

A common scenario when tokens need to be refreshed is when an app is first launched.
In this case, it is recommended to issue a heartbeat API call to a well known resource,
for example, the root folder, to ensure that tokens are refreshed before initiating an
upload or download.