Fix rotation issues with Folder Picker view controller children
BoxFolderPickerViewController is a container view controller.
One of the container VC's responsibilities is to properly size
the views contained within it.

This view controller contains a BoxFolderPickerTableViewController.
If the SDK does not have an OAuth2 session, the table view
controller is swapped it out for an authorization view
controller.

If the app is rotated while the authorization vc is active,
swapping back in the table vc when the user authenticates was
not correctly resizing the view which caused the table view
to render more narrowly than it should.

This commit follows Apple's recommendations for container view
controllers and sets the child's view bounds before making it
the active view.

@see http://developer.apple.com/library/ios/#featuredarticles/ViewControllerPGforiPhoneOS/CreatingCustomContainerViewControllers/CreatingCustomContainerViewControllers.html