Refactor BoxFoldersResourceManager to follow the style in BoxFilesResourceManager
When BoxFilesResourceManager was introduced, it bought with it some additional abstractions
for creating API operations and makes the process of creating a new endpoint more DRY.
Bring these changes to BoxFoldersResourceManager to make the SDK code more consistent.

This commit also factors out magic strings to constants for both internal and external
use.